Errors and issued addressed for PR

+++ src/server/migrations/0.6.0-0.7.0/sql/map/add_north_angle_and_circle_fraction.sql
@@ -3,5 +3,5 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 ALTER TABLE maps
-	ADD COLUMN IF NOT EXISTS north_angle FLOAT(8) DEFAULT 0.0,
-	ADD COLUMN IF NOT EXISTS max_circle_size_fraction FLOAT(8) DEFAULT 0.15;
+	ADD COLUMN IF NOT EXISTS north_angle REAL DEFAULT 0.0,
+	ADD COLUMN IF NOT EXISTS max_circle_size_fraction REAL DEFAULT 0.15;
